% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodshirt-package.R
\docType{package}
\name{goodshirt-package}
\alias{goodshirt}
\alias{goodshirt-package}
\title{goodshirt: Quotes From The TV Show The Good Place}
\description{
Get some sweet quotes from the Soul Squad.
}
\author{
\strong{Maintainer}: Adam Gruer \email{adamgruer@icloud.com}

}
\keyword{internal}
