% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quotes}
\alias{quotes}
\title{Good place quotes}
\format{
## `quotes`
A data frame with 77 rows amd 2 columns:
\describe{
\item{character}{Name of character}
\item{quotes}{Pearl of wisdom uttered by the character}}
}
\source{
<https://github.com/klaudiaroz/the-good-place-quotes>
}
\usage{
quotes
}
\description{
Some quotes from characters on the NBC show The Good Place
}
\keyword{datasets}
