% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{random_quote}
\alias{random_quote}
\title{Get A Random Quote, Benches}
\usage{
random_quote()
}
\value{
list of two elements : Character and Quote
}
\description{
Get A Random Quote, Benches
}
\examples{
random_quote
}
