% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{soul_squad}
\alias{soul_squad}
\title{Get A Random Quote, Benches}
\usage{
soul_squad()
}
\value{
list of two character vectors : character and quote
}
\description{
Get A Random Quote, Benches
}
\examples{
soul_squad()
}
